feat: replace next-auth folder
