+++ src/app/auth/signin/signInProviders.tsx
@@ -4,7 +4,7 @@ import { getProviders } from 'next-auth/react';
 
 import Kakao from './kakaoLoginButton';
 
-const SignIn = async () => {
+const SignInProviders = async () => {
   const providers = await getProviders();
   if (!providers) return <></>;
   return (
@@ -18,4 +18,4 @@ const SignIn = async () => {
   );
 };
 
-export default SignIn;
+export { SignInProviders };
